Alteração do tipo de dados de valores de Decimal para Double. Configurção da classe GetCalculateRequest para validar entrada da requisição. Ajuste no Program.cs para os testes.

+++ CalculoCDB.Server/Services/FeesService.cs
@@ -2,29 +2,29 @@
 {
 	public static class FeesService
 	{
-		private const decimal CDI = 0.009M;
-		private const decimal TB = 1.08M;
+		private const double CDI = 0.009;
+		private const double TB = 1.08;
 		private static List<Tax> Taxes = new List<Tax>
 		{
-			new() {Low =1, High = 6, Fee = 0.225M},
-			new() {Low= 7, High = 12, Fee = 0.2M },
-			new() {Low= 13, High = 24, Fee = 0.175M },
-			new() {Low= 24, High = int.MaxValue, Fee = 0.15M }
+			new() {Low =1, High = 6, Fee = 0.225},
+			new() {Low= 7, High = 12, Fee = 0.2 },
+			new() {Low= 13, High = 24, Fee = 0.175 },
+			new() {Low= 24, High = int.MaxValue, Fee = 0.15 }
 		};
 
-		public static decimal GetCdiTax()
+		public static double GetCdiTax()
 		{
 			return CDI + 1 - 1;
 		}
 
-		public static decimal GetTbTax()
+		public static double GetTbTax()
 		{
 			return TB;
 		}
 
-		public static decimal GetFeeIncome(int month)
+		public static double GetFeeIncome(int month)
 		{
-			return Taxes.Find(t => month >= t.Low && month <= t.High)?.Fee ?? 0M;
+			return Taxes.Find(t => month >= t.Low && month <= t.High)?.Fee ?? 0;
 		}
 	}
 
@@ -32,6 +32,6 @@
 	{
 		public int Low { get; set; }
 		public int High { get; set; }
-		public decimal Fee { get; set; }
+		public double Fee { get; set; }
 	}
 }
